feat: update graphql-server-mikro-orm-postgres example to MikroORM v5 and graphql-modules v1
